+++ src/i18n_check/test_frontends/search_dir/search_dir_test_file.ts
@@ -2,3 +2,6 @@
 // This file contains the following i18n keys:
 //   - "i18n._global.hello_global"
 //   - "i18n.search_dir_test_file.not_in_i18n_source_file"
+// Note: The next keys can be fixed by i18n-check --fix options on key-formatting and key-naming respectively.
+//   - "i18n.test_file.incorrectly-formatted-key"
+//   - "i18n.wrong_identifier_path.content_reference"