+++ tests/checks/test_key_naming.py
@@ -18,6 +18,7 @@ from ..test_utils import (
     fail_checks_test_file_path,
     i18n_map_fail,
     i18n_map_pass,
+    nonexistent_keys_search_dir_file,
 )
 
 invalid_name_fail = audit_invalid_i18n_key_names(
@@ -103,6 +104,18 @@ def test_invalid_key_names_check_and_fix_fail_fix_mode(capsys):
     )
     replace_text_in_file(
         path=fail_checks_test_file_path,
+        old="i18n.test_file.content_reference",
+        new="i18n.wrong_identifier_path.content_reference",
+    )
+
+    # Verify that the key in the search-dirs file has been updated appropriately.
+    with open(nonexistent_keys_search_dir_file, "r", encoding="utf-8") as f:
+        search_dir_file_content = f.read()
+
+    assert "i18n.test_file.content_reference" in search_dir_file_content
+
+    replace_text_in_file(
+        path=nonexistent_keys_search_dir_file,
         old="i18n.test_file.content_reference",
         new="i18n.wrong_identifier_path.content_reference",
     )