+++ CHANGELOG.md
@@ -10,12 +10,21 @@ See the [releases for i18n-check](https://github.com/activist-org/i18n-check/rel
 
 Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 
-## [Upcoming] i18n-check 1.19.6
+## [Upcoming] i18n-check 1.20.0
+
+### ✨ Features
+
+- Checks for `aria-label` and `alt-text` terminal punctuation was expanded to handle more languages ([#108](https://github.com/activist-org/i18n-check/issues/108)).
+- The total number of failed checks is reported to the user.
+
+### 🐛 Bug Fixes
+
+- The `key-formatting` and `key-naming` `--fix` options now also update key usage in configured `nonexistent-keys` `search-dirs` ([#118](https://github.com/activist-org/i18n-check/issues/118)).
 
 ### ♻️ Code Refactoring
 
 - The legacy Python typing system was removed in favor of the more modern approach ([#113](https://github.com/activist-org/i18n-check/issues/113)).
-- The complexity of the code was reduced to improve maintainability via checks using `complexipy` ([#115](https://github.com/activist-org/ts-backend-check/issues/115)).
+- The complexity of the code was reduced to improve maintainability via checks using `complexipy` ([#115](https://github.com/activist-org/i18n-check/issues/115)).
 
 ## i18n-check 1.19.5